Fix path parsing
Calling item/rule recusively in path/rule will cause a path like "a/b/c"
be parse as "a" "b/c", disallowing path in the nested item solves the
issue, and it causes item/rulel not reusable by path and rebol.

Because path/rule and rebol/rule could be recursively called, rename
them to nested-* respectively, and make path a wrapper instead.